Move the code for formatting push rules into a separate function
